feat: Moving the rest of the EFs in place
