+++ README.md
@@ -73,72 +73,72 @@ package tech.amikos;
 
 import tech.amikos.chromadb.*;
 import tech.amikos.chromadb.Collection;
+import tech.amikos.chromadb.embeddings.DefaultEmbeddingFunction;
 
 import java.util.*;
 
 public class Main {
-    public static void main(String[] args) {
-        try {
-            Client client = new Client(System.getenv("CHROMA_URL"));
-            client.reset();
-            EmbeddingFunction ef = new DefaultEmbeddingFunction();
-            Collection collection = client.createCollection("test-collection", null, true, ef);
-            List<Map<String, String>> metadata = new ArrayList<>();
-            metadata.add(new HashMap<String, String>() {{
-                put("type", "scientist");
-            }});
-            metadata.add(new HashMap<String, String>() {{
-                put("type", "spy");
-            }});
-            collection.add(null, metadata, Arrays.asList("Hello, my name is John. I am a Data Scientist.", "Hello, my name is Bond. I am a Spy."), Arrays.asList("1", "2"));
-            Collection.QueryResponse qr = collection.query(Arrays.asList("Who is the spy"), 10, null, null, null);
-            System.out.println(qr);
-        } catch (Exception e) {
-            System.out.println(e);
-        }
+  public static void main(String[] args) {
+    try {
+      Client client = new Client(System.getenv("CHROMA_URL"));
+      client.reset();
+      EmbeddingFunction ef = new DefaultEmbeddingFunction();
+      Collection collection = client.createCollection("test-collection", null, true, ef);
+      List<Map<String, String>> metadata = new ArrayList<>();
+      metadata.add(new HashMap<String, String>() {{
+        put("type", "scientist");
+      }});
+      metadata.add(new HashMap<String, String>() {{
+        put("type", "spy");
+      }});
+      collection.add(null, metadata, Arrays.asList("Hello, my name is John. I am a Data Scientist.", "Hello, my name is Bond. I am a Spy."), Arrays.asList("1", "2"));
+      Collection.QueryResponse qr = collection.query(Arrays.asList("Who is the spy"), 10, null, null, null);
+      System.out.println(qr);
+    } catch (Exception e) {
+      System.out.println(e);
     }
+  }
 }
 ```
 
 ### Example OpenAI Embedding Function
 
-In this example we rely on `tech.amikos.chromadb.OpenAIEmbeddingFunction` to generate embeddings for our documents.
+In this example we rely on `tech.amikos.chromadb.embeddings.openai.OpenAIEmbeddingFunction` to generate embeddings for our documents.
 
 | **Important**: Ensure you have `OPENAI_API_KEY` environment variable set
 
 ```java
 package tech.amikos;
 
-import com.google.gson.internal.LinkedTreeMap;
 import tech.amikos.chromadb.Client;
 import tech.amikos.chromadb.Collection;
 import tech.amikos.chromadb.EmbeddingFunction;
-import tech.amikos.chromadb.OpenAIEmbeddingFunction;
+import tech.amikos.chromadb.embeddings.openai.OpenAIEmbeddingFunction;
 
 import java.util.*;
 
 public class Main {
-    public static void main(String[] args) {
-        try {
-            Client client = new Client(System.getenv("CHROMA_URL"));
-            String apiKey = System.getenv("OPENAI_API_KEY");
-            EmbeddingFunction ef = new OpenAIEmbeddingFunction(apiKey, "text-embedding-3-small");
-            Collection collection = client.createCollection("test-collection", null, true, ef);
-            List<Map<String, String>> metadata = new ArrayList<>();
-            metadata.add(new HashMap<String, String>() {{
-                put("type", "scientist");
-            }});
-            metadata.add(new HashMap<String, String>() {{
-                put("type", "spy");
-            }});
-            collection.add(null, metadata, Arrays.asList("Hello, my name is John. I am a Data Scientist.", "Hello, my name is Bond. I am a Spy."), Arrays.asList("1", "2"));
-            Collection.QueryResponse qr = collection.query(Arrays.asList("Who is the spy"), 10, null, null, null);
-            System.out.println(qr);
-        } catch (Exception e) {
-            e.printStackTrace();
-            System.out.println(e);
-        }
+  public static void main(String[] args) {
+    try {
+      Client client = new Client(System.getenv("CHROMA_URL"));
+      String apiKey = System.getenv("OPENAI_API_KEY");
+      EmbeddingFunction ef = new OpenAIEmbeddingFunction(apiKey, "text-embedding-3-small");
+      Collection collection = client.createCollection("test-collection", null, true, ef);
+      List<Map<String, String>> metadata = new ArrayList<>();
+      metadata.add(new HashMap<String, String>() {{
+        put("type", "scientist");
+      }});
+      metadata.add(new HashMap<String, String>() {{
+        put("type", "spy");
+      }});
+      collection.add(null, metadata, Arrays.asList("Hello, my name is John. I am a Data Scientist.", "Hello, my name is Bond. I am a Spy."), Arrays.asList("1", "2"));
+      Collection.QueryResponse qr = collection.query(Arrays.asList("Who is the spy"), 10, null, null, null);
+      System.out.println(qr);
+    } catch (Exception e) {
+      e.printStackTrace();
+      System.out.println(e);
     }
+  }
 }
 ```
 
@@ -221,7 +221,7 @@ The above should output:
 
 ### Example Hugging Face Sentence Transformers Embedding Function
 
-In this example we rely on `tech.amikos.chromadb.HuggingFaceEmbeddingFunction` to generate embeddings for our documents.
+In this example we rely on `tech.amikos.chromadb.embeddings.hf.HuggingFaceEmbeddingFunction` to generate embeddings for our documents.
 
 | **Important**: Ensure you have `HF_API_KEY` environment variable set
 
@@ -230,31 +230,32 @@ package tech.amikos;
 
 import tech.amikos.chromadb.*;
 import tech.amikos.chromadb.Collection;
+import tech.amikos.chromadb.embeddings.hf.HuggingFaceEmbeddingFunction;
 
 import java.util.*;
 
 public class Main {
-    public static void main(String[] args) {
-        try {
-            Client client = new Client(System.getenv("CHROMA_URL"));
-            client.reset();
-            String apiKey = System.getenv("HF_API_KEY");
-            EmbeddingFunction ef = new HuggingFaceEmbeddingFunction(apiKey);
-            Collection collection = client.createCollection("test-collection", null, true, ef);
-            List<Map<String, String>> metadata = new ArrayList<>();
-            metadata.add(new HashMap<String, String>() {{
-                put("type", "scientist");
-            }});
-            metadata.add(new HashMap<String, String>() {{
-                put("type", "spy");
-            }});
-            collection.add(null, metadata, Arrays.asList("Hello, my name is John. I am a Data Scientist.", "Hello, my name is Bond. I am a Spy."), Arrays.asList("1", "2"));
-            Collection.QueryResponse qr = collection.query(Arrays.asList("Who is the spy"), 10, null, null, null);
-            System.out.println(qr);
-        } catch (Exception e) {
-            System.out.println(e);
-        }
+  public static void main(String[] args) {
+    try {
+      Client client = new Client(System.getenv("CHROMA_URL"));
+      client.reset();
+      String apiKey = System.getenv("HF_API_KEY");
+      EmbeddingFunction ef = new HuggingFaceEmbeddingFunction(apiKey);
+      Collection collection = client.createCollection("test-collection", null, true, ef);
+      List<Map<String, String>> metadata = new ArrayList<>();
+      metadata.add(new HashMap<String, String>() {{
+        put("type", "scientist");
+      }});
+      metadata.add(new HashMap<String, String>() {{
+        put("type", "spy");
+      }});
+      collection.add(null, metadata, Arrays.asList("Hello, my name is John. I am a Data Scientist.", "Hello, my name is Bond. I am a Spy."), Arrays.asList("1", "2"));
+      Collection.QueryResponse qr = collection.query(Arrays.asList("Who is the spy"), 10, null, null, null);
+      System.out.println(qr);
+    } catch (Exception e) {
+      System.out.println(e);
     }
+  }
 }
 ```
