+++ src/main/java/tech/amikos/chromadb/embeddings/hf/CreateEmbeddingResponse.java
@@ -1,4 +1,4 @@
-package tech.amikos.hf;
+package tech.amikos.chromadb.embeddings.hf;
 
 import com.google.gson.Gson;
