+++ src/main/java/tech/amikos/chromadb/embeddings/hf/HuggingFaceClient.java
@@ -1,4 +1,4 @@
-package tech.amikos.hf;
+package tech.amikos.chromadb.embeddings.hf;
 
 import com.google.gson.Gson;
 import okhttp3.*;
