+++ src/main/java/tech/amikos/chromadb/embeddings/hf/HuggingFaceEmbeddingFunction.java
@@ -1,31 +1,27 @@
-package tech.amikos.chromadb;
+package tech.amikos.chromadb.embeddings.hf;
 
 
-import tech.amikos.hf.CreateEmbeddingRequest;
-import tech.amikos.hf.CreateEmbeddingResponse;
-import tech.amikos.hf.HuggingFaceClient;
+import tech.amikos.chromadb.EmbeddingFunction;
 
 import java.util.List;
 
 public class HuggingFaceEmbeddingFunction implements EmbeddingFunction {
-
     private final String hfAPIKey;
-
+    private final HuggingFaceClient client;
     public HuggingFaceEmbeddingFunction(String hfAPIKey) {
         this.hfAPIKey = hfAPIKey;
-
+         this.client = new HuggingFaceClient(this.hfAPIKey);
     }
 
     @Override
     public List<List<Float>> createEmbedding(List<String> documents) {
-        HuggingFaceClient client = new HuggingFaceClient(this.hfAPIKey);
-        CreateEmbeddingResponse response = client.createEmbedding(new CreateEmbeddingRequest().inputs(documents.toArray(new String[0])));
+
+        CreateEmbeddingResponse response = this.client.createEmbedding(new CreateEmbeddingRequest().inputs(documents.toArray(new String[0])));
         return response.getEmbeddings();
     }
 
     @Override
     public List<List<Float>> createEmbedding(List<String> documents, String model) {
-        HuggingFaceClient client = new HuggingFaceClient(this.hfAPIKey);
         client.modelId(model);
         CreateEmbeddingResponse response = client.createEmbedding(new CreateEmbeddingRequest().inputs(documents.toArray(new String[0])));
         return response.getEmbeddings();
