+++ src/main/java/tech/amikos/chromadb/embeddings/openai/CreateEmbeddingResponse.java
@@ -1,4 +1,4 @@
-package tech.amikos.openai;
+package tech.amikos.chromadb.embeddings.openai;
 
 import com.google.gson.Gson;
 import com.google.gson.annotations.SerializedName;
