+++ src/test/java/tech/amikos/chromadb/Utils.java
@@ -1,3 +1,5 @@
+package tech.amikos.chromadb;
+
 import java.io.BufferedReader;
 import java.io.File;
 import java.io.FileReader;
