+++ src/test/java/tech/amikos/chromadb/embeddings/TestDefaultEmbeddings.java
@@ -1,14 +1,15 @@
-package tech.amikos.chromadb;
+package tech.amikos.chromadb.embeddings;
 
 import org.apache.commons.io.FileUtils;
 import org.junit.Test;
+import tech.amikos.chromadb.EmbeddingFunction;
 
 import java.util.Arrays;
 import java.util.List;
 
 import static org.junit.Assert.*;
 
-public class TestDefaultEmbeddingFunction {
+public class TestDefaultEmbeddings {
     // this represents the output of sentence-transformers/all-MiniLM-L6-v2 for "Hello, my name is John. I am a Data Scientist.", "Hello, I am Jane and I am an ML researcher."
     private static float[][] groundThruth = {{-0.09585458785295486f,
             0.00948028638958931f,
