+++ src/test/java/tech/amikos/chromadb/embeddings/ollama/TestOllamaEmbeddings.java
@@ -13,7 +13,7 @@ import java.util.List;
 
 import static org.junit.Assert.*;
 
-public class TestOllamaEmbeddingFunction {
+public class TestOllamaEmbeddings {
     static GenericContainer ollamaContainer;
 
     @BeforeClass
